x/issuecoin/types/message_mintcoin_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"issuecoin/testutil/sample"
)

func TestMsgMintcoin_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgMintcoin
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgMintcoin{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgMintcoin{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
